+++ config/routes.rb
@@ -78,6 +78,7 @@ Rails.application.routes.draw do
       get     "/secrets"                            => 'secrets#batch'
       get     "/edge/secrets/:account"              => 'edge#all_secrets'
       get     "/edge/hosts/:account"                => 'edge#all_hosts'
+      get     "/edge/edges/:account"                => 'edge#all_edges'
       get     "/edge/slosilo_keys/:account"         => 'edge#slosilo_keys'
 
       put     "/policies/:account/:kind/*identifier" => 'policies#put'