+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Pull Slosilo library to Conjur
 - Change Slosilo id from "authn:account:host/user" to "authn:account:host/user:current"
 - Add update slosilo key option to slosilo put key function
+- Mask Edge IP https://ca-il-jira.il.cyber-ark.com:8443/browse/ONYX-38139
 
 ## [1.0.0-cloud] - 2023-06-07
 ### Changed